services/alert_service/requirements.txt
fastapi
uvicorn
pydantic
redis
requests
python-telegram-bot
pymongo
python-dotenv